+++ CHANGELOG.md
@@ -23,6 +23,14 @@ A feature release on top of `1.4.0`, backwards compatible.
   defines how special entries (e.g., symlinks) are to be handled when
   adding directories to CAS.
 
+- `just serve` accepts a new subkey `"client address"` for the key
+  `"execution endpoint"` in the configuration file. It informs the
+  `serve` instance that the client will access the remote-execution
+  endpoint via the given `"client address"` and not using the one
+  listed in the subkey `"address"`. This feature allows to position
+  `just serve` next to the remote-execution endpoint behind some
+  form of address translation.
+
 ### Fixes
 
 - `just-import-git` now correctly inherits pragmas for imported